reporter-web-app: Refactor SummaryView into multiple subcomponents
Splitting SummaryView to keep the code manageable as we
keep on adding more functionality.
